[TASK] Re-introduce design document migrations
It is now possible again to import design documents into CouchDB instances.

Example:
You have 2 database instances configured, 'main' and 'log', and you
have 1 design document named 'my-global-designdocument' that has
to be added to both databases, and 1 design document named 'log' that
only has to be added to the log database.

For that you add the following structure to your package folder:

`-- Migrations
    `-- CouchDB
        `-- DesignDocuments
            |-- _all
            |   `-- my-global-designdocument
            |       `-- views
            |           `-- my-view
            |               |-- map.js
            |               `-- reduce.js
            `-- log
                `-- log
                    `-- views
                        `-- my-log-view
                            |-- map.js
                            `-- reduce.js

This change also removes the requirement of having a 'default' instance.