[TASK] Depend on Doctrine CouchDB ODM commit in master
This dependency is made as recent master of Doctrine CouchDB ODM
pulls in Doctrine instantiator which will prevent the constructor
of models being called.
This will break dependency inject in Flow.

We could not lock to the last released version as that version
did not yet have a composer manifest. For this reason we locked
to commit ce00bf3 which is the last commit before the dependency
to Doctrine instantiator was introduced.